# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaPhoto API
  description: |-
    This OpenAPI document describes the functionalities of Wasaphoto, a social media platform. The website has all the basic features such as banning/unbanning other users, liking/unliking photos, upload/delete photos and commenting/uncommenting on photos. 
    
    This API supports Cross-Origin Resource Sharing (CORS), allowing it to interact with other domains in this API for data reasons to increase the website's usability.
    
    This API was created in 2023/2024.
    
  version: "10"
paths:
...
  /photo/upload/{userID}:
    post:
      tags: ["photo"]
      summary: Upload a photo
      description: |-
        This endpoint allows a user to upload a photo.
      security:
        - bearerAuth: []     
      operationId: uploadPhoto
      parameters:
        - in: path
          name: userID
          required: true
          description: |-
            post photo on user's account by identifying them with their userID.
          schema:
            type: string
      requestBody:
        description: |-
          new photo file upload request. If successful, a new photoID is
          created. Each photo has a unique identifier which is the photoID.
        content:
          multipart/form-data:
              schema: 
                type: object
                properties:
                  file:
                    type: string
                    format: binary
                    
          
        required: true

      responses:
        '200':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
...
  /photo/{photoID}/remove:
    delete:
      tags: ["photo"]
      summary: Delete a Photo
      description: |-
        This endpoint allows a user to delete a photo.
      security:
        - bearerAuth: []
      operationId: deletePhoto
      parameters:
        - in: path
          name: photoID
          required: true
          description: |-
            PhotoID is a unique identifier. This is needed to retrieve the
            specific photo to delete.
          schema:
            type: string
            example: "pic1"
      responses:
        '200':
          description: Photo is deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "photo removed successfully"
                
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "photoID not found"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        App session token value. This value is sent by the server in the login
        and session refresh APIs.
  schemas:
...
    Photo:
      type: object
      description: A photo
      properties:
        photoID: 
          type: integer
          description: the unique identifier for a photo
          minLength: 3
          maxLength: 500
          example: 1
        userID:
          type: string
          description: user identifier
          minLength: 10
          maxLength: 80
        photo:
          type: string
          format: binary
          minLength: 1
          maxLength: 100000
          
        like_count:
          type: integer
          description: The number of likes of the photo.
          example: 22
        comment_count:
          type: integer
          description: The number of comments of a photo.
          example: 22
        date:
          type: string
          format: date-time
          minLength: 20
          maxLength: 21
          description: the date and time when the photo is uploaded.
          pattern: '^.*?$'
          example: 2021-05-10T06:11:09Z
        username:
          type: string
          example: Maria
          pattern: 'ˆ.*?$'
          minLength: 3
          maxLength: 16
          description: the username
        isLiked:
          type: boolean
          example: true
          description: check whether the current user has liked the photo
...